allow editing things
Unfortunately, I've run into the age-old commit-is-slow problem -- but
it's absolutely necessary to commit after a delete in order to get
things to work (because SQLAlchemy, I suppose). The tags go haywire
otherwise. I don't think it's that big of a deal, but it is kind of
irritating. I also saved a substantial amount of time, for whatever
reason, or at least my perception is that I did, by iterating over long
lists of tags twice to update the tag list rather than just clearing it
and adding it again. Maybe addItem() is slower than I thought.

Also, things work now! Yay! Now we just have to be able to add items and
this can begin to be somewhat useful.