implement tag renaming
To do next: a merge function, because renaming doesn't merge correctly,
and it's almost certainly better to be explicit about what you want to
do here rather than end up accidentally merging a tag in. Even if we get
a good undo function working, the user might well not notice what she's
done until it's too late to undo it and the database is nastily screwed
up.