implement private bookmarks
One useful tweak would be to preclude the setting of the 'private'
checkbox if private bookmarks are not shown -- currently, doing so
causes the bookmark to become deselected and disappear as soon as the
changes are saved.

Also, actually protecting the setting via a password or similar would
probably be useful / helpful.