fix description text box settings
This is now a QPlainTextEdit; before it accepted rich text if you pasted
it in but you couldn't do anything useful to edit it since there wasn't
a formatting toolbar, which was really annoying.

Also, hitting tab now adjust focus instead of inserting a literal tab,
which I can't see being very useful.