clean up main file
This removes the majority of the pylint issues and does a little bit of
refactoring in the model file. More will be necessary to really clean
things up, but this is a start (I hacked this out and never refactored
at all because I got it working well enough for my purposes before I got
to that point, plus I was a college freshman and didn't know much about
design!).