uniquify names of bookmarks
Super confusing to have several with the same name, plus could cause
problems in the link check routine now.

Also added indexes on the names, since that's a very common search. The
filter remains unindexed because it does contains searches; if
performance becomes a concern we might be able to tap into SQLite's
full-text search options.